Add/Update: Added detail page for pokemon. Updated existing pages and CSS for same.
