+++ pokemon/src/App.css
@@ -2,13 +2,23 @@
   width: 40%;
     display: grid;
     padding: 2%;
-    grid-template-columns: repeat(auto-fit,minmax(300px, 1fr));
-    /* border-width: 2px; */
     box-shadow: 0px 2px;
     border-style: solid;
     border-color: coral;
+    margin:2%;
 }
 
 .card_layout {
   display:inline-flex;
+}
+
+.btn {
+  margin: 2%;
+  background-color: rgba(144,206,241,0.3);
+  color:#000;
+  box-shadow: 0px 0px 1px 1px #90CEF1;
+}
+
+.align {
+  text-align: center;
 }
